refactor: update import paths for consistency and improve asset references
